requirements.txt
feedparser==5.2.1
python-vlc
